Moved test functions from azure provider to tests_common package for reusability in other providers (#45478)
* refactor: Moved helper functions which allows you to run deferrable operators in unit test from azure to test_common utils so it can be reused across other providers.

* refactor: Reformatted xcom_pull method from mock_context

* refactor: Removed unused imports from conftest

* refactor: Removed unused import from conftest

---------
