+++ providers/tests/microsoft/conftest.py
@@ -21,11 +21,10 @@ import json
 import random
 import re
 import string
-from collections.abc import Iterable
 from inspect import currentframe
 from json import JSONDecodeError
 from os.path import dirname, join
-from typing import TYPE_CHECKING, Any, TypeVar
+from typing import Any, TypeVar
 from unittest.mock import MagicMock
 
 import pytest
@@ -34,15 +33,6 @@ from msgraph_core import APIVersion
 
 from airflow.models import Connection
 from airflow.providers.microsoft.azure.hooks.powerbi import PowerBIHook
-
-try:
-    from airflow.sdk.definitions.context import Context
-except ImportError:
-    # TODO: Remove once provider drops support for Airflow 2
-    from airflow.utils.context import Context
-
-if TYPE_CHECKING:
-    from sqlalchemy.orm import Session
 
 T = TypeVar("T", dict, str, Connection)
 
@@ -115,52 +105,6 @@ def mock_response(status_code, content: Any = None, headers: dict | None = None)
     response.content = content
     response.json.side_effect = JSONDecodeError("", "", 0)
     return response
-
-
-def mock_context(task) -> Context:
-    from airflow.models import TaskInstance
-    from airflow.utils.session import NEW_SESSION
-    from airflow.utils.state import TaskInstanceState
-    from airflow.utils.xcom import XCOM_RETURN_KEY
-
-    values: dict[str, Any] = {}
-
-    class MockedTaskInstance(TaskInstance):
-        def __init__(
-            self,
-            task,
-            run_id: str | None = "run_id",
-            state: str | None = TaskInstanceState.RUNNING,
-            map_index: int = -1,
-        ):
-            super().__init__(task=task, run_id=run_id, state=state, map_index=map_index)
-            self.values: dict[str, Any] = {}
-
-        def xcom_pull(
-            self,
-            task_ids: str | Iterable[str] | None = None,
-            dag_id: str | None = None,
-            key: str = XCOM_RETURN_KEY,
-            include_prior_dates: bool = False,
-            session: Session = NEW_SESSION,
-            *,
-            map_indexes: int | Iterable[int] | None = None,
-            default: Any = None,
-            run_id: str | None = None,
-        ) -> Any:
-            if map_indexes:
-                return values.get(
-                    f"{task_ids or self.task_id}_{dag_id or self.dag_id}_{key}_{map_indexes}", default
-                )
-            return values.get(f"{task_ids or self.task_id}_{dag_id or self.dag_id}_{key}", default)
-
-        def xcom_push(self, key: str, value: Any, session: Session = NEW_SESSION, **kwargs) -> None:
-            values[f"{self.task_id}_{self.dag_id}_{key}_{self.map_index}"] = value
-
-    values["ti"] = MockedTaskInstance(task=task)
-
-    # See https://github.com/python/mypy/issues/8890 - mypy does not support passing typed dict to TypedDict
-    return Context(values)  # type: ignore[misc]
 
 
 def remove_license_header(content: str) -> str:
